Move TlvReader toAny method to ChipTool
