+++ examples/android/CHIPTool/app/src/main/java/com/google/chip/chiptool/clusterclient/MultiAdminClientFragment.kt
@@ -16,7 +16,6 @@ import chip.devicecontroller.model.ChipAttributePath
 import chip.devicecontroller.model.ChipEventPath
 import chip.devicecontroller.model.InvokeElement
 import chip.devicecontroller.model.NodeState
-import chip.jsontlv.toAny
 import chip.tlv.AnonymousTag
 import chip.tlv.TlvReader
 import chip.tlv.TlvWriter
@@ -24,6 +23,7 @@ import com.google.chip.chiptool.ChipClient
 import com.google.chip.chiptool.GenericChipDeviceListener
 import com.google.chip.chiptool.R
 import com.google.chip.chiptool.databinding.MultiAdminClientFragmentBinding
+import com.google.chip.chiptool.util.toAny
 import kotlinx.coroutines.*
 
 class MultiAdminClientFragment : Fragment() {
